requirements.txt
aiohttp==3.6.2
aiomultiprocess==0.7.0
aniso8601==8.0.0
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
Brotli==1.0.7
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cloudscraper==1.2.16
colorama==0.4.1
cryptography==2.7
deen==2.0.1
defusedxml==0.6.0
dicttoxml==1.7.4
dnspython==1.16.0
docopt==0.6.2
ecdsa==0.15
EditorConfig==0.12.2
et-xmlfile==1.0.1
exrex==0.10.5
fire==0.2.1
Flask==1.1.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
fuzzywuzzy==0.18.0
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
Js2Py==0.66
jsbeautifier==1.10.3
loguru==0.4.0
lxml==4.4.2
MarkupPy==1.14
MarkupSafe==1.1.1
multidict==4.7.2
odfpy==1.4.0
openpyxl==2.4.11
Pillow==7.0.0
pipreqs==0.4.10
psutil==5.7.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycparser==2.19
pyjsparser==2.7.1
pyOpenSSL==18.0.0
PyQt5==5.14.1
PyQt5-sip==12.7.1
PySocks==1.7.1
python-jose==3.1.0
python-nmap==0.6.1
pytz==2019.3
PyYAML==5.2
records==0.5.3
requests==2.22.0
requests-file==1.4.3
requests-toolbelt==0.9.1
rsa==4.0
shodan==1.21.3
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
tablib==0.14.0
tenacity==6.1.0
termcolor==1.1.0
tld==0.9.3
tldextract==2.2.2
tqdm==4.41.0
tzlocal==2.0.0
urllib3==1.25.3
Werkzeug==1.0.0
win32-setctime==1.0.1
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
xmltodict==0.12.0
yarg==0.1.9
yarl==1.4.2
